Add BLDSS request and response building
This commit adds building of the BLDSS API request and subsequent
parsing and return of the response into our standard schema.

Search is very naive at the moment, we only support:

title=<something>

this is transformed into a general index search at the BLDSS end. We
need to add the ability to specify indexes for searching.