Add "Action" endpoint and authentication
This commit adds an "/action" endpoint that exists to allow an arbitrary
action to be performed. Currently this consists of a POST to
/api/orders, but will be abstracted to make it able to perform all
actions.

This commit also adds a new BLDSSAuth class which provides the ability
to generate an authentication header. A new BLDSSRequest class has also
been created that provides the ability to make a request to the API.