Switch to sync requests and fix RA ID
Although ISO18626 requests are asynchronous, we don't have to
overcomplicate things for ourselves and use BLDSS async requests. Here
we are specifying the use of synchronous requests. This will obviously
mean we need to emulate an async SA to the RA but, on reflection, that's
going to be waaay less confusing than using the BLDSS async flow.

This commit also fixes the formation of the RA ID we pass to BLDSS as
customerRef, it should be our local request ID