Add ill-connector-info to list of abilities
When we delare our abilities, we should include the ability to provide
info.

We really could do with this ability list being produced from some
introspection mechanism, rather than a manually maintained list, but
I've no idea how we'd do that without. It's available in
ModuleDescriptor, but I'm not aware of any mechanism that gives us
access to those properties